Server/task_remainder/requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
Django==1.9.4
django-filter==1.0.4
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
kombu==4.0.2
pkg-resources==0.0.0
PyJWT==1.5.0
pytz==2017.2
redis==2.10.5
vine==1.1.3
